【改修】回転角度のカラム定義を整数から小数へ修正

+++ database/migrations/2021_12_08_025548_create_media_figures_table.php
@@ -22,7 +22,7 @@ class CreateMediaFiguresTable extends Migration
             $table->integer('top');
             $table->integer('width');
             $table->integer('height');
-            $table->integer('degree')->default(1);
+            $table->float('degree', 4, 1)->default(1);
             $table->float('global_alpha', 3, 2)->default(1);
             $table->integer('layer');
             $table->boolean('is_draw_fill');
